+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed
+### Added
+- Ability to decode nested `arrays`
 
+### Changed
+* BREAKING: The Decoding for `Variants` was not returning the decoded value type in the `json` representation. Now `Variants` would be decoded like `{"field":["uint32",100]}`
 * BREAKING: The serialization for `ExtendedAsset` was aligned with the `eos` codebase.  Beforehand, it would serialize the field name `"Contract"` with a capital `C`, and the `Asset` field as `"asset"` where it should have been `"quantity"`.